Set path when updating the Antora content catalog
People directly using the Antora content catalog (it's a long story :) ) may need the path set for consistency with the rest of the entries in there.

I'm not sure if it makes more sense to pass the full path around with the image, but in the interest of changing the least code, I've just added the path when setting the file in the content catalog.